+++ package/network/utils/owipcalc/Makefile
@@ -7,17 +7,17 @@
 include $(TOPDIR)/rules.mk
 
 PKG_NAME:=owipcalc
-PKG_RELEASE:=5
+PKG_RELEASE:=6
 PKG_LICENSE:=Apache-2.0
 
 include $(INCLUDE_DIR)/package.mk
-
 
 define Package/owipcalc
   SECTION:=utils
   CATEGORY:=Utilities
   TITLE:=Simple IPv4/IPv6 address calculator
   MAINTAINER:=Jo-Philipp Wich <jo@mein.io>
+  DEPENDS:=+libowipcalc
 endef
 
 define Package/owipcalc/description
@@ -27,18 +27,45 @@ define Package/owipcalc/description
   for dhcp pools.
 endef
 
+define Package/libowipcalc
+  SECTION:=libs
+  CATEGORY:=Libraries
+  TITLE:=Owipcalc Library
+endef
+
+define Package/libowipcalc/description
+  The owipcalc library supports a number of calculations and tests to work
+  with ip-address ranges, this is useful for programms that e.g. need to
+  partition ipv6-prefixes into small subnets or to calculate address ranges
+  for dhcp pools.
+endef
+
 define Build/Configure
 endef
 
 define Build/Compile
 	$(TARGET_CC) $(TARGET_CFLAGS) \
-		-o $(PKG_BUILD_DIR)/owipcalc $(PKG_BUILD_DIR)/owipcalc.c
+		-shared -fPIC -o $(PKG_BUILD_DIR)/libowipcalc.so $(PKG_BUILD_DIR)/owipcalc.c
+	$(TARGET_CC) $(TARGET_CFLAGS) \
+		-o $(PKG_BUILD_DIR)/owipcalc $(PKG_BUILD_DIR)/main.c -L$(PKG_BUILD_DIR)/ -lowipcalc
 endef
 
+define Build/InstallDev
+	$(INSTALL_DIR) $(1)/usr/include/
+	$(CP) $(PKG_BUILD_DIR)/owipcalc.h $(1)/usr/include/
+	$(INSTALL_DIR) $(1)/usr/lib
+	$(INSTALL_BIN) $(PKG_BUILD_DIR)/libowipcalc.so $(1)/usr/lib/libowipcalc.so
+endef
 
 define Package/owipcalc/install
 	$(INSTALL_DIR) $(1)/usr/bin
 	$(INSTALL_BIN) $(PKG_BUILD_DIR)/owipcalc $(1)/usr/bin/owipcalc
 endef
 
+define Package/libowipcalc/install
+	$(INSTALL_DIR) $(1)/usr/lib
+	$(INSTALL_BIN) $(PKG_BUILD_DIR)/libowipcalc.so $(1)/usr/lib/libowipcalc.so
+endef
+
+$(eval $(call BuildPackage,libowipcalc))
 $(eval $(call BuildPackage,owipcalc))
